+++ src/apps/dfx/ethereum/dfx.staking.contract-position-fetcher.ts
@@ -1,94 +1,40 @@
-import { Inject } from '@nestjs/common';
-import _ from 'lodash';
+import { PositionTemplate } from '~app-toolkit/decorators/position-template.decorator';
 
-import { IAppToolkit, APP_TOOLKIT } from '~app-toolkit/app-toolkit.interface';
-import { Register } from '~app-toolkit/decorators';
-import { buildDollarDisplayItem } from '~app-toolkit/helpers/presentation/display-item.present';
-import { getImagesFromToken, getLabelFromToken } from '~app-toolkit/helpers/presentation/image.present';
-import { ContractType } from '~position/contract.interface';
-import { PositionFetcher } from '~position/position-fetcher.interface';
-import { ContractPosition } from '~position/position.interface';
-import { claimable, supplied } from '~position/position.utils';
-import { Network } from '~types/network.interface';
+import { DfxStakingContractPositionFetcher } from '../common/dfx.staking.contract-position-fetcher';
 
-import { Addresses } from '../addresses';
-import { DfxContractFactory } from '../contracts';
-import { DFX_DEFINITION } from '../dfx.definition';
-
-const appId = DFX_DEFINITION.id;
-const groupId = DFX_DEFINITION.groups.staking.id;
-const network = Network.ETHEREUM_MAINNET;
-
-@Register.ContractPositionFetcher({ appId, groupId, network })
-export class EthereumDfxStakingContractPositionFetcher implements PositionFetcher<ContractPosition> {
-  constructor(
-    @Inject(APP_TOOLKIT) private readonly appToolkit: IAppToolkit,
-    @Inject(DfxContractFactory) private readonly dfxContractFactory: DfxContractFactory,
-  ) {}
-
-  async getPositions() {
-    const stakingDefinitions = Addresses[network].amm.map(({ staking: stakingAddress, curve: curveAddress }) => ({
-      address: stakingAddress,
-      stakedTokenAddress: curveAddress,
-      rewardTokenAddress: Addresses[network].dfx,
-    }));
-
-    // Reward token is DFX which is a base token
-    const baseTokens = await this.appToolkit.getBaseTokenPrices(network);
-    // Staked tokens are AMM LPs so resolve these
-    const appTokens = await this.appToolkit.getAppTokenPositions({
-      appId: DFX_DEFINITION.id,
-      groupIds: [DFX_DEFINITION.groups.dfxCurve.id],
-      network,
-    });
-    const allTokens = [...baseTokens, ...appTokens];
-
-    // Create a multicall wrapper instance to batch chain RPC calls together
-    const multicall = this.appToolkit.getMulticall(network);
-
-    const positions = await Promise.all(
-      stakingDefinitions.map(async ({ address, stakedTokenAddress, rewardTokenAddress }) => {
-        const stakedToken = allTokens.find(v => v.address === stakedTokenAddress);
-        const rewardToken = allTokens.find(v => v.address === rewardTokenAddress);
-        if (!stakedToken || !rewardToken) return null;
-
-        const tokens = [supplied(stakedToken), claimable(rewardToken)];
-
-        // Instantiate a smart contract instance pointing to the jar token address
-        const contract = this.dfxContractFactory.dfxCurve({ address: stakedToken.address, network });
-
-        // Request the jar token balance of this farm
-        const [balanceRaw] = await Promise.all([multicall.wrap(contract).balanceOf(address)]);
-
-        // Denormalize the balance as the TVL
-        const liquidity = Number(balanceRaw) / 10 ** stakedToken.decimals;
-
-        // Prepare display props
-        const label = getLabelFromToken(stakedToken);
-        const images = tokens.map(v => getImagesFromToken(v)).flat();
-        const secondaryLabel = buildDollarDisplayItem(stakedToken.price);
-        const statsItems = [{ label: 'Liquidity', value: buildDollarDisplayItem(liquidity) }];
-
-        const position: ContractPosition = {
-          type: ContractType.POSITION,
-          appId,
-          groupId,
-          address,
-          network,
-          tokens,
-          dataProps: {
-            liquidity,
-          },
-          displayProps: {
-            label,
-            secondaryLabel,
-            images,
-            statsItems,
-          },
-        };
-        return position;
-      }),
-    );
-    return _.compact(positions);
-  }
+@PositionTemplate()
+export class EthereumDfxStakingContractPositionFetcher extends DfxStakingContractPositionFetcher {
+  groupLabel = 'DFX Staking';
+  stakingDefinitions = [
+    {
+      address: '0x84bf8151394dcf32146965753b28760550f3d7a8',
+      stakedTokenAddress: '0xa6c0cbcaebd93ad3c6c94412ec06aaa37870216d', // cadc-Usdc
+      rewardTokenAddress: '0x888888435fde8e7d4c54cab67f206e4199454c60',
+    },
+    {
+      address: '0x5eaaeff69f2ab64d1cc0244fb31b236ca989544f',
+      stakedTokenAddress: '0x1a4ffe0dcbdb4d551cfca61a5626afd190731347', // eurs-Usdc
+      rewardTokenAddress: '0x888888435fde8e7d4c54cab67f206e4199454c60',
+    },
+    {
+      address: '0xd52d48db08e8224ef6e2be8f54f3c84e790b1c32',
+      stakedTokenAddress: '0x2bab29a12a9527a179da88f422cdaaa223a90bd5', // xsgd-Usdc
+      rewardTokenAddress: '0x888888435fde8e7d4c54cab67f206e4199454c60',
+    },
+    {
+      address: '0xe06fa52e0d2d58fe192285bfa0507f09cdd9824a',
+      stakedTokenAddress: '0xe9669516e09f5710023566458f329cce6437aaac', // nzds-Usdc
+      rewardTokenAddress: '0x888888435fde8e7d4c54cab67f206e4199454c60',
+    },
+    {
+      address: '0xddb720069fdfe7be2e2883a1c06be0f353f7c4c8',
+      stakedTokenAddress: '0xc574a613a3900e4314da13eb2287f13689a5b64d', // tryb-Usdc
+      rewardTokenAddress: '0x888888435fde8e7d4c54cab67f206e4199454c60',
+    },
+    {
+      address: '0xe29b7285c1169a9765e2a9bfe74209077bee55d6',
+      stakedTokenAddress: '0xdd39379ab7c93b9baae29e6ec03795d0bc99a889', // xidr-Usdc
+      rewardTokenAddress: '0x888888435fde8e7d4c54cab67f206e4199454c60',
+    },
+  ];
 }