+++ src/apps/trisolaris/index.ts
@@ -1,3 +1,0 @@
-export { TRISOLARIS_DEFINITION, TrisolarisAppDefinition } from './trisolaris.definition';
-export { TrisolarisAppModule } from './trisolaris.module';
-export { TrisolarisContractFactory } from './contracts';